requirements.txt
# Core Deep Learning Frameworks
tensorflow>=2.8.0
keras>=2.8.0

# Numerical Computing
numpy>=1.21.0
scipy>=1.7.0

# Data Processing
pandas>=1.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Image Processing
opencv-python>=4.5.0
Pillow>=8.3.0
scikit-image>=0.18.0

# Machine Learning Utilities
scikit-learn>=1.0.0

# Configuration Management
pyyaml>=5.4.0

# Progress Bars
tqdm>=4.62.0

# Logging and Monitoring
tensorboard>=2.8.0

# Utilities
python-dotenv>=0.19.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Code Quality
black>=21.0
flake8>=3.9.0
isort>=5.9.0

# Jupyter Notebooks
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.6.0
